% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ds18b20.R
\docType{class}
\name{Ds18b20-class}
\alias{Ds18b20-class}
\title{Ds18b20 Class Representation.}
\description{
Ds18b20 Class Representation.
}
\section{Slots}{

\describe{
\item{\code{device_file}}{The sensor's configuration file.}
}}

