% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ds18b20_sensor.R
\name{get_ds18b20_sensor}
\alias{get_ds18b20_sensor}
\title{Get a DS18B20 sensor instance.}
\usage{
get_ds18b20_sensor(sensor_id = NA)
}
\arguments{
\item{sensor_id}{A character of length one with the ID of the sensor to use,
as found in the `/sys/bus/w1/devices/` directory. If `NA`, it will return
the first sensor found. This parameter should be used only if more than one
DS18B20 sensor is connected. Defaults to `NA`.}
}
\value{
A ds18b20 S3 object, this object will be set to work with the selected
  sensor.
}
\description{
Returns a DS18B20 sensor instance ready to get its temperature.
The sensor should be set to use with the Raspberry's 1-Wire protocol.
}
\examples{
\dontrun{
# If just one sensor connected, this would be the normal use case.
sensor <- get_ds18b20_sensor()

# If two sensors connected, say "28-0300a2790676" and "28-0300a0881918".
sensor_1 <- get_ds18b20_sensor("28-0300a2790676")
sensor_2 <- get_ds18b20_sensor("28-0300a0881918")
}

}
