% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature.R
\name{get_temperature.ds18b20}
\alias{get_temperature.ds18b20}
\title{Get the temperature of a DS18B20 sensor instance.}
\usage{
\method{get_temperature}{ds18b20}(x, scale = "celsius", max_retries = 10, ...)
}
\arguments{
\item{x}{A sensor instance.}

\item{scale}{One of `"celsius"`, `"fahrenheit"` or `"kelvin"`, indicating
the desired return scale.}

\item{max_retries}{A numeric of length one indicating how many times it
should retry in case of issues.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Reads the temperature of a given sensor instance.
}
\examples{
\dontrun{
# Get a DS18B20 sensor instance.
sensor <- get_ds18b20_sensor()
# Read the temperature in Celsius and in Fahrenheit.
get_temperature(sensor)
get_temperature(sensor, "fahrenheit")
}

}
