% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature.R
\name{get_temperature}
\alias{get_temperature}
\title{Get the temperature of a sensor instance.}
\usage{
get_temperature(x, ...)
}
\arguments{
\item{x}{A sensor instance.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Reads the temperature of a given sensor instance.
}
\examples{
\dontrun{
# Get a DS18B20 sensor instance.
sensor <- get_ds18b20_sensor()
# Read the temperature.
get_temperature(sensor)
}

}
